test(scheduling): fold contributed-channel tests into runner-service suite (realness ratchet: no new IAgentRuntime casts)
